add AI Stage Unit

+++ DragonMarbleDomain/Contents.cs
@@ -15,7 +15,7 @@ namespace DragonMarble
 	public static class RandomUtil
 	{
 		public static float Next(float min, float max){
-			return RandomFactory.NewRandom().Next(min * 1000, max * 1000) / 1000f;
+            return RandomFactory.NewRandom().Next((int)(min * 1000), (int)max * 1000) / 1000f;
 		}
 	}
